Parameterize the patterns used to match notifications to their handlers.
I have preexisting SNS topics that I just want to listen in on, and I didn't want to have to create new ones just to use this out of the box.

I got fed up with having outstanding changes that pulling it out was probably a good idea, especially considering assertible/lambda-cloudwatch-slack#16.